+++ src/parsers/ansi/ansi_commands.rs
@@ -411,7 +411,7 @@ impl Parser {
         let (top, bottom) = match self.parsed_numbers.len() {
             2 => (self.parsed_numbers[0] - 1, self.parsed_numbers[1] - 1),
             1 => (0, self.parsed_numbers[0] - 1),
-            0 => (0, buf.terminal_state.get_height() as i32),
+            0 => (0, buf.terminal_state.get_height()),
             _ => {
                 return Err(Box::new(ParserError::UnsupportedEscapeSequence(
                     self.current_escape_sequence.clone(),
@@ -442,7 +442,7 @@ impl Parser {
         let (left, right) = match self.parsed_numbers.len() {
             2 => (self.parsed_numbers[0] - 1, self.parsed_numbers[1] - 1),
             1 => (0, self.parsed_numbers[0] - 1),
-            0 => (0, buf.terminal_state.get_height() as i32),
+            0 => (0, buf.terminal_state.get_height()),
             _ => {
                 return Err(Box::new(ParserError::UnsupportedEscapeSequence(
                     self.current_escape_sequence.clone(),
@@ -483,7 +483,7 @@ impl Parser {
                 self.parsed_numbers[0] - 1,
                 self.parsed_numbers[1] - 1,
                 self.parsed_numbers[2] - 1,
-                buf.get_width() as i32,
+                buf.get_width(),
             ),
             4 => (
                 self.parsed_numbers[0] - 1,
@@ -535,7 +535,7 @@ impl Parser {
                 let bottom = if let Some((_, b)) = buf.terminal_state.get_margins_top_bottom() {
                     b
                 } else {
-                    buf.get_height() as i32 - 1
+                    buf.get_height() - 1
                 };
                 buf.terminal_state.set_margins_top_bottom(n, bottom);
             }
@@ -551,7 +551,7 @@ impl Parser {
                 let right = if let Some((_, r)) = buf.terminal_state.get_margins_left_right() {
                     r
                 } else {
-                    buf.get_width() as i32 - 1
+                    buf.get_width() - 1
                 };
                 buf.terminal_state.set_margins_left_right(n, right);
             }
@@ -879,13 +879,7 @@ impl Parser {
         let pl = self.parsed_numbers[3];
         let pb = self.parsed_numbers[4];
         let pr = self.parsed_numbers[5];
-        if pt > pb
-            || pl > pr
-            || pr > buf.get_width() as i32
-            || pb > buf.get_height() as i32
-            || pl < 0
-            || pt < 0
-        {
+        if pt > pb || pl > pr || pr > buf.get_width() || pb > buf.get_height() || pl < 0 || pt < 0 {
             return Err(Box::new(ParserError::UnsupportedEscapeSequence(format!(
                 "invalid area for requesting checksum pt:{pt} pl:{pl} pb:{pb} pr:{pr}"
             ))));
@@ -1061,8 +1055,8 @@ impl Parser {
                         self.current_escape_sequence
                     ))));
                 }
-                let width = self.parsed_numbers[2].min(132).max(1) as usize;
-                let height = self.parsed_numbers[1].min(60).max(1) as usize;
+                let width = self.parsed_numbers[2].min(132).max(1);
+                let height = self.parsed_numbers[1].min(60).max(1);
                 buf.terminal_state.set_width(width);
                 buf.terminal_state.set_height(height);
                 Ok(CallbackAction::ResizeTerminal(width, height))
@@ -1121,21 +1115,15 @@ impl Parser {
     fn get_rect_area(&self, buf: &Buffer, offset: usize) -> (i32, i32, i32, i32) {
         let top_line: i32 = self.parsed_numbers[offset]
             .max(1)
-            .min(buf.get_line_count().max(buf.get_height()) as i32)
+            .min(buf.get_line_count().max(buf.get_height()))
             - 1;
-        let left_column = self.parsed_numbers[offset + 1]
-            .max(1)
-            .min(buf.get_width() as i32)
-            - 1;
+        let left_column = self.parsed_numbers[offset + 1].max(1).min(buf.get_width()) - 1;
 
         let bottom_line = self.parsed_numbers[offset + 2]
             .max(1)
-            .min(buf.get_line_count().max(buf.get_height()) as i32)
+            .min(buf.get_line_count().max(buf.get_height()))
             - 1;
-        let right_column = self.parsed_numbers[offset + 3]
-            .max(1)
-            .min(buf.get_width() as i32)
-            - 1;
+        let right_column = self.parsed_numbers[offset + 3].max(1).min(buf.get_width()) - 1;
 
         (top_line, left_column, bottom_line, right_column)
     }
